# config-api/docs/swagger.yaml
basePath: /api
...
host: localhost:3001
info:
  contact:
    name: Beedevz
    url: https://github.com/beedevz/hive
  description: Config-driven homelab start page backend API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Hive API
  version: "1.0"
paths:
...
  /adapters/{type}:
    get:
      description: Fetches live stats from the named adapter for the given service.
        Results are cached for 60 s.
      parameters:
      - description: Adapter type (e.g. pihole, proxmox, sonarr)
        in: path
        name: type
        required: true
        type: string
      - description: Service name as defined in config
        in: query
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AdapterResult with stats array
          schema:
            type: object
        "400":
          description: Missing adapter type or service parameter
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Run adapter for a service
      tags:
      - adapters
...
  /config:
    get:
      consumes:
      - application/json
      description: GET returns the full configuration as JSON. PUT replaces it (creates
        a backup first).
      parameters:
      - description: Config object (PUT only)
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Config object or {"status":"ok"}
          schema:
            type: object
        "400":
          description: Invalid JSON
          schema:
            type: string
        "401":
          description: Unauthorized (PUT)
          schema:
            type: string
        "404":
          description: Config not found (GET)
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get or save config
      tags:
      - config
    put:
      consumes:
      - application/json
      description: GET returns the full configuration as JSON. PUT replaces it (creates
        a backup first).
      parameters:
      - description: Config object (PUT only)
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Config object or {"status":"ok"}
          schema:
            type: object
        "400":
          description: Invalid JSON
          schema:
            type: string
        "401":
          description: Unauthorized (PUT)
          schema:
            type: string
        "404":
          description: Config not found (GET)
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get or save config
      tags:
      - config
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by your Hive token.
    in: header
    name: Authorization
    type: apiKey
  HiveToken:
    in: header
    name: X-Hive-Token
    type: apiKey
swagger: "2.0"